+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Unreleased
 
+- Add support for `host.name` in Resource attributes
 - Add support for baggage propagation (#51)
 - Updating dependencies to 0.14b0 (#51)
 - Add support for tracecontext propagation (#52)